Refactored block table with structure field #3323 #3324
